+++ climada_petals/hazard/tc_rainfield.py
@@ -229,7 +229,7 @@ class TCRain(Hazard):
     def from_tracks(
         cls,
         tracks: TCTracks,
-        centroids: Optional[Centroids] = None,
+        centroids: Centroids = None,
         pool: Optional[pathos.pools.ProcessPool] = None,
         model: str = 'R-CLIPER',
         model_kwargs: Optional[dict] = None,
@@ -298,7 +298,8 @@ class TCRain(Hazard):
         tracks : climada.hazard.TCTracks
             Tracks of storm events.
         centroids : Centroids, optional
-            Centroids where to model TC. Default: global centroids at 360 arc-seconds resolution.
+            Centroids where to model TC. Default: centroids at 360 arc-seconds resolution within
+            tracks' bounds.
         pool : pathos.pool, optional
             Pool that will be used for parallel computation of rain fields. Default: None
         model : str, optional
@@ -398,20 +399,15 @@ class TCRain(Hazard):
         """
         num_tracks = tracks.size
         if centroids is None:
-            centroids = Centroids.from_base_grid(res_as=360, land=True)
-
-        if not centroids.coord.size:
-            centroids.set_meta_to_lat_lon()
+            centroids = Centroids.from_pnt_bounds(tracks.get_bounds(), res=0.1)
 
         if ignore_distance_to_coast:
             # Select centroids with lat <= max_latitude
             [idx_centr_filter] = (np.abs(centroids.lat) <= max_latitude).nonzero()
         else:
             # Select centroids which are inside max_dist_inland_km and lat <= max_latitude
-            if not centroids.dist_coast.size:
-                centroids.set_dist_coast()
             [idx_centr_filter] = (
-                (centroids.dist_coast <= max_dist_inland_km * 1000)
+                (centroids.get_dist_coast() <= max_dist_inland_km * 1000)
                 & (np.abs(centroids.lat) <= max_latitude)
             ).nonzero()